# server/.api/apis/dcd/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Smartproxy",
    "version": "1.2"
  },
  "servers": [
    {
      "url": "https://api.decodo.com"
    }
  ],
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "Authorization"
      }
    }
  },
  "security": [
    {
      "sec0": []
    }
  ],
  "paths": {
    "/v2/endpoints/{type}": {
      "get": {
        "summary": "Get endpoints by type",
        "description": "Get a list of endpoints by a specified type for Residential subscription",
        "operationId": "get-endpoints-by-type",
        "parameters": [
          {
            "name": "type",
            "in": "path",
            "description": "Possible parameters: random, sticky",
            "schema": {
              "type": "string"
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "200",
            "content": {
              "application/json": {
                "examples": {
                  "Result": {
                    "value": "[\n  0:{\n    \"location\":\"Random\"\n    \"hostname\":\"gate.smartproxy.com\"\n    \"port_range\":\"7000\"\n  }\n  1:{\n    \"location\":\"Canada\"\n    \"hostname\":\"ca.smartproxy.com\"\n    \"port_range\":\"20000\"\n}"
                  }
                }
              }
            }
          },
          "400": {
            "description": "400",
            "content": {
              "application/json": {
                "examples": {
                  "Result": {
                    "value": "{\n  \"error_code\":\"bad_request\"\n  \"message\":\"Bad proxy type provided.\"\n}"
                  }
                }
              }
            }
          }
        },
        "deprecated": false,
        "security": []
      }
    },
    "/v2/endpoints": {
      "get": {
        "summary": "Get endpoints",
        "description": "Get types of endpoints for Residential subscription",
        "operationId": "get-endpoints",
        "responses": {
          "200": {
            "description": "200",
            "content": {
              "application/json": {
                "examples": {
                  "Result": {
                    "value": "[\n  0:{\n    \"type\":\"random\"\n    \"available_locations\":\"148\"\n    \"url\":\"endpoints/random\"\n  }\n  1:{\n    \"type\":\"sticky\"\n    \"available_locations\":\"62\"\n    \"url\":\"endpoints/sticky\"\n  }\n]"
                  }
                }
              }
            }
          },
          "400": {
            "description": "400",
            "content": {
              "application/json": {
                "examples": {
                  "Result": {
                    "value": "{}"
                  }
                },
                "schema": {
                  "type": "object",
                  "properties": {}
                }
              }
            }
          }
        },
        "deprecated": false,
        "security": []
      }
    },
...
    "/v2/sub-users": {
      "get": {
        "summary": "Get sub users",
        "description": "Get a list of active sub users for Residential subscription",
        "operationId": "get-sub-users",
        "parameters": [
          {
            "name": "service_type",
            "in": "query",
            "description": "Possible parameters: residential_proxies",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "200",
            "content": {
              "application/json": {
                "examples": {
                  "Result": {
                    "value": "[\n  0:{\n    \"id\":24570,\n    \"username\":\"name\",\n    \"status\":\"active\",\n    \"created_at\":\"2019-03-23 13:08\",\n    \"traffic\" : 1.13,\n    \"traffic_limit\":NULL,\n    \"service_type\":\"residential_proxies\",\n    \"traffic_count_from\":\"2019-03-23 13:10\",\n    \"auto_disable\":false\n  }\n]"
                  }
                }
              }
            }
          },
          "400": {
            "description": "400",
            "content": {
              "application/json": {
                "examples": {
                  "Result": {
                    "value": "{\n  \"error_code\":\"not_found\"\n  \"message\":\"Resource not found\"\n}"
                  }
                }
              }
            }
          }
        },
        "deprecated": false,
        "security": [
          {
            "ApiKeyAuth": []
          }
        ]
      },
      "post": {
        "summary": "Create sub user",
        "description": "Create a new sub user for Residential subscription",
        "operationId": "create-sub-user",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "username",
                  "password",
                  "service_type"
                ],
                "properties": {
                  "username": {
                    "type": "string",
                    "description": "From 6 to 64 characters long. Only letters, numbers and underscores allowed.",
                    "format": "json"
                  },
                  "password": {
                    "type": "string",
                    "description": "9 or more characters. Must include at least one upper case letter and a number. Symbols @ and : are not allowed",
                    "format": "json"
                  },
                  "service_type": {
                    "type": "string",
                    "description": "Possible parameters: residential_proxies, shared_proxies",
                    "default": "residential_proxies"
                  },
                  "traffic_limit": {
                    "type": "number",
                    "format": "float"
                  },
                  "auto_disable": {
                    "type": "boolean",
                    "description": "Disables subuser when the traffic limit is hit",
                    "default": false
                  },
                  "traffic_count_from": {
                    "type": "string",
                    "description": "Handles subuser subscription traffic starting from the specified date. Use date format yyyy-mm-dd hh:mm:ss.",
                    "format": "date"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "201",
            "content": {
              "application/json": {
                "examples": {
                  "Result": {
                    "value": "{}"
                  }
                },
                "schema": {
                  "type": "object",
                  "properties": {}
                }
              }
            }
          },
          "400": {
            "description": "400",
            "content": {
              "application/json": {
                "examples": {
                  "Result": {
                    "value": "{\n  \"error_code\": \"bad_request\",\n  \"message\": \"Can not process request\",\n  \"error\": {\n    \"username\": \"username must be between 3 and 64 characters. Only numbers, characters and underscore are allowed\",\n    \"password\": \"password must be between 9 and 64 characters and must contain at least one number. Symbols @ and : are invalid\"\n  }\n}"
                  }
                },
                "schema": {
                  "type": "object",
                  "properties": {
                    "error_code": {
                      "type": "string",
                      "example": "bad_request"
                    },
                    "message": {
                      "type": "string",
                      "example": "Can not process request"
                    },
                    "error": {
                      "type": "object",
                      "properties": {
                        "username": {
                          "type": "string",
                          "example": "username must be between 3 and 64 characters. Only numbers, characters and underscore are allowed"
                        },
                        "password": {
                          "type": "string",
                          "example": "password must be between 9 and 64 characters and must contain at least one number. Symbols @ and : are invalid"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "deprecated": false
      }
    },
    "/v2/sub-users/{sub_user_id}": {
      "get": {
        "summary": "Get sub user",
        "description": "Get a single active sub user",
        "operationId": "get-sub-user",
        "parameters": [
          {
            "name": "sub_user_id",
            "in": "path",
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "200",
            "content": {
              "application/json": {
                "examples": {
                  "Result": {
                    "value": "{\n  \"id\": 1,\n  \"username\": \"name\",\n  \"status\": \"active\",\n  \"created_at\": \"2021-07-01 00:00\",\n  \"traffic\": 0,\n  \"traffic_limit\": null,\n  \"service_type\": \"type\",\n  \"traffic_count_from\": null,\n  \"auto_disable\": false\n}"
                  }
                },
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "integer",
                      "example": 1,
                      "default": 0
                    },
                    "username": {
                      "type": "string",
                      "example": "name"
                    },
                    "status": {
                      "type": "string",
                      "example": "active"
                    },
                    "created_at": {
                      "type": "string",
                      "example": "2021-07-01 00:00"
                    },
                    "traffic": {
                      "type": "integer",
                      "example": 0,
                      "default": 0
                    },
                    "traffic_limit": {},
                    "service_type": {
                      "type": "string",
                      "example": "type"
                    },
                    "traffic_count_from": {},
                    "auto_disable": {
                      "type": "boolean",
                      "example": false,
                      "default": true
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "400",
            "content": {
              "application/json": {
                "examples": {
                  "Result": {
                    "value": "{\n  \"error_code\": \"bad_request\",\n  \"message\": \"Can not process request\"\n}"
                  }
                },
                "schema": {
                  "type": "object",
                  "properties": {
                    "error_code": {
                      "type": "string",
                      "example": "bad_request"
                    },
                    "message": {
                      "type": "string",
                      "example": "Can not process request"
                    }
                  }
                }
              }
            }
          }
        },
        "deprecated": false
      },
      "put": {
        "summary": "Update sub user",
        "description": "Update password or traffic limit of specified sub user",
        "operationId": "update-sub-user",
        "parameters": [
          {
            "name": "sub_user_id",
            "in": "path",
            "description": "\"Get sub users\" received sub user ID",
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "required": true
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "traffic_limit": {
                    "type": "number",
                    "description": "Traffic limit",
                    "format": "float"
                  },
                  "auto_disable": {
                    "type": "boolean",
                    "description": "Disables subuser when the traffic limit is hit",
                    "default": false
                  },
                  "password": {
                    "type": "string",
                    "format": "json"
                  },
                  "traffic_count_from": {
                    "type": "string",
                    "description": "Handles subuser subscription traffic starting from the specified date. Use date format yyyy-mm-dd hh:mm:ss.",
                    "format": "date"
                  },
                  "status": {
                    "type": "string",
                    "description": "Possible values: 'active' or 'disabled'"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "201",
            "content": {
              "application/json": {
                "examples": {
                  "Result": {
                    "value": "{}"
                  }
                },
                "schema": {
                  "type": "object",
                  "properties": {}
                }
              }
            }
          },
          "400": {
            "description": "400",
            "content": {
              "application/json": {
                "examples": {
                  "Result": {
                    "value": "{\n  \"error_code\": \"bad_request\",\n  \"message\": \"Can not process request\",\n  \"error\": {\n    \"password\": \"password must be between 9 and 64 characters and must contain at least one number. Symbols @ and : are invalid\"\n  }\n}"
                  }
                },
                "schema": {
                  "type": "object",
                  "properties": {
                    "error_code": {
                      "type": "string",
                      "example": "bad_request"
                    },
                    "message": {
                      "type": "string",
                      "example": "Can not process request"
                    },
                    "error": {
                      "type": "object",
                      "properties": {
                        "password": {
                          "type": "string",
                          "example": "password must be between 9 and 64 characters and must contain at least one number. Symbols @ and : are invalid"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "deprecated": false
      },
      "delete": {
        "summary": "Delete sub user",
        "description": "Delete specified sub user",
        "operationId": "delete-sub-user",
        "parameters": [
          {
            "name": "sub_user_id",
            "in": "path",
            "description": "\"Get sub users\" received sub user ID",
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "required": true
          }
        ],
        "responses": {
          "204": {
            "description": "204",
            "content": {
              "application/json": {
                "examples": {
                  "Result": {
                    "value": "{}"
                  }
                },
                "schema": {
                  "type": "object",
                  "properties": {}
                }
              }
            }
          },
          "400": {
            "description": "400",
            "content": {
              "application/json": {
                "examples": {
                  "Result": {
                    "value": "{\n  \"error_code\":\"not_found\"\n  \"message\":\"Resource not found\"\n}"
                  }
                }
              }
            }
          }
        },
        "deprecated": false
      }
    },
...
    "/v2/whitelisted-ips": {
      "get": {
        "summary": "Get whitelisted IPs",
        "description": "Get a list of whitelisted IPs for Residential subscription",
        "operationId": "get-whitelisted-ips",
        "responses": {
          "200": {
            "description": "200",
            "content": {
              "application/json": {
                "examples": {
                  "Result": {
                    "value": "[\n  0:{\n    \"id\":69489\n    \"ip\":\"127.0.0.1\"\n    \"enabled\":true\n    \"created_at\":\"2019-05-22 12:44\"\n\t}\n]"
                  }
                }
              }
            }
          },
          "400": {
            "description": "400",
            "content": {
              "application/json": {
                "examples": {
                  "Result": {
                    "value": "{\n  \"error_code\":\"not_found\"\n  \"message\":\"Resource not found\"\n}"
                  }
                }
              }
            }
          }
        },
        "deprecated": false
      },
      "post": {
        "summary": "Add whitelisted IPs",
        "description": "Whitelist your IPs here for Residential subscription",
        "operationId": "add-whitelisted-ips",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "IPAddressList": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "200",
            "content": {
              "application/json": {
                "examples": {
                  "Result": {
                    "value": "{}"
                  }
                },
                "schema": {
                  "type": "object",
                  "properties": {}
                }
              }
            }
          },
          "400": {
            "description": "400",
            "content": {
              "application/json": {
                "examples": {
                  "Result": {
                    "value": "{\n  \"error_code\":\"bad_request\"\n  \"message\":\"Can not process request\"\n}"
                  }
                }
              }
            }
          }
        },
        "deprecated": false
      }
    },
...
  },
...
}